+++ src/app/game/regular/boxscore/page.tsx
@@ -1,7 +1,7 @@
 import { Banner } from '@/components/common/banner'
 import TabMenu from '@/components/common/tab-menu'
 import Breadcrumbs from '@/components/tailwind-ui/breadcrumbs/simple-with-chevrons'
-import { ScoreBoard } from '@/components/boxscore/score-board'
+import { ScoreBoard } from '@/components/boxscore/score-board/score-board'
 
 import { GAME_BANNER_DATA } from '@/contants/index'
