+++ src/components/boxscore/score-board/label.tsx
